chore: encapsulate determination of presence of touchscreen into usewindowdimensions hook
